% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTProx_glasso.R
\name{BTProx_glasso}
\alias{BTProx_glasso}
\title{Backtracking Proximal Gradient Method for Graphical Lasso}
\usage{
BTProx_glasso(S, n, lambda, warm.start = FALSE, init.Omega = NULL)
}
\arguments{
\item{S}{Sample covariance matrix}

\item{n}{Sample size}

\item{lambda}{Penalty parameter}

\item{warm.start}{Bool indicator for whether to do warm start for estimation}

\item{init.Omega}{Initial value for inverse covariance matrix, only needed when warm.start=TRUE}
}
\value{
Estimated inverse covariance matrix
}
\description{
This function uses backtracking to search for step size in proximal gradient method and solve for inverse covariance matrix
}
\examples{
n=10
p=5
set.seed(42)
samples=Generate_AR1_pair(n,p,rho=0.7,cov_AR1 = TRUE)
S=cov(samples$X)/n*(n-1)
BT_est=BTProx_glasso(S,n,lambda = 0.1)
}
