% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_AR1_pair.R
\name{Generate_AR1_pair}
\alias{Generate_AR1_pair}
\title{Simulate data points with AR1 covariance structure}
\usage{
Generate_AR1_pair(n, p, rho, cov_AR1 = TRUE)
}
\arguments{
\item{n}{Sample size}

\item{p}{Sample dimension}

\item{rho}{Coefficient of AR1 model}

\item{cov_AR1}{Bool indicator for whether the covariance or inverse covariance has AR1 structure}
}
\value{
A list containing true covariance, inverse covariance and simulated data
}
\description{
This function simulate data points follow multivariate normal distribution with given dimension and structure
}
\examples{
n=10
p=5
set.seed(42)
samples=Generate_AR1_pair(n,p,rho=0.7,cov_AR1 = TRUE)
}
