% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newton_Prox_primal_glasso.R
\name{Newton_Prox_primal_glasso}
\alias{Newton_Prox_primal_glasso}
\title{Self-Concordant Proximal Newton Method(primal) for Graphical Lasso}
\usage{
Newton_Prox_primal_glasso(S, n, lambda, warm.start = F, init.Omega = NULL)
}
\arguments{
\item{S}{Sample covariance matrix}

\item{n}{Sample size}

\item{lambda}{Penalty parameter}

\item{warm.start}{Bool indicator for whether to do warm start for estimation}

\item{init.Omega}{Initial value for inverse covariance matrix, only needed when warm.start=TRUE}
}
\value{
Estimated inverse covariance matrix
}
\description{
This function uses local Hessian estimate and primal solution to search for direction and step size in proximal gradient method and solve for inverse covariance matrix
}
